Caite #4
"Hotpep does this thing, whereby if there are no CAZyme hits found in a genome by Hotpep/PPR, it will not create any file. What happens is that the whole of run_dbcan.py ends up terminating when it cannot find Results/output.txt when it wants to copy it to the output directory as Hotpep.out. When I ran 75,000 genomes, this happened for several thousand genomes and I had to go back and redo them. We fixed this in the previous implementation by going into ‘parallel_group_many_proteins….py’ and commenting out the line that is the last if statement (if len(hit_array) > 0:) and just running that block of code regardless of the length of the hit_array variable, so that it will at least output an empty file. "

-Thank you! Great suggestions!