Make installable using setuptools, for upload to pip and then conda
Resolves #18.

All scripts made executable, with interpreter shebang, so they can be
run in-place without explicitly invoking python interpreter.  The
point being, invoking the python interpreter explicitly means that the
script will not be found along the path, which breaks when this
package is installed properly.

Remove explicit invocations of python from Hotpep scripts.

Fix a bug creating Hotpep subdir if parent directory didn't exist.
mkdir -p would also have worked, but we don't need to spawn a
subprocess to make a directory.

Ensure Hotpep can find its data files, both when installed and when
running in place in development mode.

Add a run script for development mode, run_dev_dbcan, which is
required to set PATH and PYTHONPATH so this can be run in place,
without being installed.