Fixes and conda packaging for run_dbcan (#101)
* Edited hmmscan_parser and run_dbcan so that they can be imported as python modules and called from other python scripts without invoking subprocesses, etc. They retain the oriignal functionality of being called via CLI through a shell.

Additionally, prediction.py had a bug in get_validation_results that caused an empty folder structure to be created when the output was given as an absolute path instead of relative path. This bug should now be fixed.

* Missed some bugfix updates from the main branch of run_dbcan in last commit, which are included here.

Also changed the call to hmmscan_parser in runHmmScan to a function call instead of subprocess. This will remove the overhead of creating a new subprocess, and makes more sense given that hmmscan_parser is now a module import.

* Added build files for conda installation.

Also moved around the arg input in run_dbcan so that the entry points in setup.py is correct.

* Added note about changing git_url to meta.yaml when merging into main repo.

* Fixed bug where including branch in git_url prevented conda build from executing successfully.

* Added missing dependencies to conda build file.

* Changed run_dbcan imports to solve module not found error in conda installation.

* Removed bioconda as a package. It's a channel, not a package.

* Added init file for proper module installation, and licence file name to meta file.

* made setuptools a build dependency instead of run dependency

* Edited import to function properly

* Edited import to function properly

* Changed to noarch in meta.yaml

* Change meta.yaml to point to main repo for merging into it.

Also pull version info from setup.py, meaning there is a single source of truth for version number now.